test transitive reduction on perlmutter
