+++ src/main.cpp
@@ -59,7 +59,9 @@ constexpr int root = 0; /* root process rank */
 int parse_cli(int argc, char *argv[]);
 void print_kmer_histogram(const KmerCountMap& kmermap, std::shared_ptr<CommGrid> commgrid);
 void parallel_write_paf(const CT<Overlap>::PSpParMat& R, DistributedFastaData& dfd, char const *pafname);
-std::string getpafname();
+CT<uint64_t>::PDistVec find_contained_reads(const CT<Overlap>::PSpParMat& R);
+std::string get_overlap_paf_name();
+std::string get_string_paf_name();
 
 int main(int argc, char **argv)
 {
@@ -289,31 +291,18 @@ int main(int argc, char **argv)
         R = PairwiseAlignment(dfd, *B, mat, mis, gap, xdrop_cutoff);
         timer.stop_and_log("pairwise alignment");
 
-        CT<int>::PSpParMat containedQmat = R->Prune([](const Overlap& o) { return !o.containedQ; }, false);
-        CT<int>::PSpParMat containedTmat = R->Prune([](const Overlap& o) { return !o.containedT; }, false);
-
-        CT<int>::PDistVec containedQvec = containedQmat.Reduce(Row, std::logical_or<int>(), 0);
-        CT<int>::PDistVec containedTvec = containedTmat.Reduce(Column, std::logical_or<int>(), 0);
-
-        CT<int>::PDistVec contained = containedQvec;
-        contained.EWiseOut(containedTvec, std::logical_or<int>(), contained);
-
-        contained.DebugPrint();
-        CT<uint64_t>::PDistVec toerase = contained.FindInds([](const int& v) { return v > 0; });
-        toerase.DebugPrint();
-
-        R->PruneFull(toerase, toerase);
-
         elbalog.log_overlap_matrix(*R);
-
-        parallel_write_paf(*R, dfd, getpafname().c_str());
+        parallel_write_paf(*R, dfd, get_overlap_paf_name().c_str());
 
         timer.start();
+        auto contained = find_contained_reads(*R);
+        R->PruneFull(contained, contained);
+
         S = TransitiveReduction(*R);
-        timer.stop_and_log("transitive reduction");
+        timer.stop_and_log("contained read removal and transitive reduction");
         R.reset();
 
-        parallel_write_paf(*S, dfd, "string.paf");
+        parallel_write_paf(*S, dfd, get_string_paf_name().c_str());
 
         walltimer.stop_and_log("wallclock");
 
@@ -503,9 +492,27 @@ void parallel_write_paf(const CT<Overlap>::PSpParMat& R, DistributedFastaData& d
     MPI_File_close(&fh);
 }
 
-std::string getpafname()
+CT<uint64_t>::PDistVec find_contained_reads(const CT<Overlap>::PSpParMat& R)
+{
+    CT<int>::PSpParMat containedQmat = const_cast<CT<Overlap>::PSpParMat&>(R).Prune([](const Overlap& o) { return !o.containedQ; }, false);
+    CT<int>::PSpParMat containedTmat = const_cast<CT<Overlap>::PSpParMat&>(R).Prune([](const Overlap& o) { return !o.containedT; }, false);
+    CT<int>::PDistVec containedQvec = containedQmat.Reduce(Row, std::logical_or<int>(), 0);
+    CT<int>::PDistVec containedTvec = containedTmat.Reduce(Column, std::logical_or<int>(), 0);
+    CT<int>::PDistVec contained = containedQvec;
+    contained.EWiseOut(containedTvec, std::logical_or<int>(), contained);
+    return contained.FindInds([](const int& v) { return v > 0; });
+}
+
+std::string get_string_paf_name()
 {
     std::ostringstream ss;
-    ss << output_prefix << ".paf";
+    ss << output_prefix << ".string.paf";
+    return ss.str();
+}
+
+std::string get_overlap_paf_name()
+{
+    std::ostringstream ss;
+    ss << output_prefix << ".overlap.paf";
     return ss.str();
 }
